tracing/pyproject.toml
[project]
name = "tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "braintrust>=0.1.2",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
